+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @layerswap/utils
 
+## 2.3.0
+
+### Patch Changes
+
+- Updated dependencies [95d95a1]
+- Updated dependencies [31d0d1f]
+- Updated dependencies [5bd03c7]
+- Updated dependencies [bdf46a0]
+  - @layerswap/widget-types@2.3.0
+
+## 2.2.0
+
+### Minor Changes
+
+- Global improvements and bug fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerswap/widget-types@2.2.0
+
 ## 2.1.0
 
 ### Minor Changes