+++ packages/wallets/adapters/imtblPassport/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @layerswap/wallet-imtbl-passport
 
+## 2.3.0
+
+### Patch Changes
+
+- Updated dependencies [bdf46a0]
+- Updated dependencies [f26c1e7]
+  - @layerswap/wallet-core@2.3.0
+
+## 2.2.0
+
+### Minor Changes
+
+- Global improvements and bug fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerswap/wallet-core@2.2.0
+
 ## 2.1.0
 
 ### Minor Changes