+++ packages/wallets/all/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @layerswap/wallets
 
+## 2.3.0
+
+### Patch Changes
+
+- Updated dependencies [dc962d6]
+- Updated dependencies [95d95a1]
+- Updated dependencies [31d0d1f]
+- Updated dependencies [5bd03c7]
+- Updated dependencies [bdf46a0]
+- Updated dependencies [f26c1e7]
+- Updated dependencies [bdf46a0]
+  - @layerswap/wallet-starknet@2.3.0
+  - @layerswap/wallet-paradex@2.3.0
+  - @layerswap/wallet-stellar@2.3.0
+  - @layerswap/wallet-fuel@2.3.0
+  - @layerswap/wallet-ton@2.3.0
+  - @layerswap/wallet-tron@2.3.0
+  - @layerswap/wallet-bitcoin@2.3.0
+  - @layerswap/wallet-svm@2.3.0
+  - @layerswap/widget-types@2.3.0
+  - @layerswap/wallet-core@2.3.0
+  - @layerswap/wallet-evm@2.3.0
+  - @layerswap/wallet-imtbl-passport@2.3.0
+  - @layerswap/utils@2.3.0
+
+## 2.2.0
+
+### Minor Changes
+
+- Global improvements and bug fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @layerswap/utils@2.2.0
+  - @layerswap/wallet-bitcoin@2.2.0
+  - @layerswap/wallet-evm@2.2.0
+  - @layerswap/wallet-fuel@2.2.0
+  - @layerswap/wallet-imtbl-passport@2.2.0
+  - @layerswap/wallet-paradex@2.2.0
+  - @layerswap/wallet-starknet@2.2.0
+  - @layerswap/wallet-stellar@2.2.0
+  - @layerswap/wallet-svm@2.2.0
+  - @layerswap/wallet-ton@2.2.0
+  - @layerswap/wallet-tron@2.2.0
+  - @layerswap/wallet-core@2.2.0
+  - @layerswap/widget-types@2.2.0
+
 ## 2.1.0
 
 ### Minor Changes